+++ src/assets/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.1.6 (2021-11-23)
+copy paste function 
+
+using short codes for copy paste
+
 ## 0.1.5 (2021-11-22)
 
 Using google form for feedback now